Fix context usage to be more secure. Python 3.11+ now working
