+++ environment.yml
@@ -18,5 +18,3 @@ dependencies:
   - python-json-logger
   - dill
   - pandas
-  - pip:
-    - line-profiler
