+++ tests/conftest.py
@@ -19,15 +19,8 @@ pytest_plugins=[
     'fixtures.command_fixtures', 'fixtures.chunk_fixtures',
     'fixtures.western_fixtures', 'fixtures.data_fixtures',
     'fixtures.cli_fixtures', 'fixtures.fusion_fixtures',
-    'fixtures.metric_fixtures']
-
-@pytest.fixture(scope="session", autouse=True)
-def configure_dask() -> None:
-    dask.config.set(scheduler="single-threaded")
-
-@pytest.fixture(scope="function")
-def threaded_dask() -> None:
-    dask.config.set(scheduler="threads")
+    'fixtures.metric_fixtures', 'fixtures.dask_fixtures'
+]
 
 @pytest.fixture(scope='function')
 def tdb_filepath(storage_config) -> Generator[str, None, None]:
